[DEV-172] moved chatbot services to own chatbot subfolder
