+++ app/services/chatbot/static-pages/de.json
