+++ app/services/chatbot/static-pages/fr.json
